requirements.txt
einops
evo
hydra-core
lightning
rich
scikit-learn
torchtyping
wandb